Lispy is a superset of Paredit, if you configure it like this:
(lispy-set-key-theme '(special paredit c-digits))
In addition to Paredit functionality, Lispy also has:

refactoring
REPL integration for Elisp, Clojure, Common Lisp, Scheme, Hy etc.
integration with avy and ivy
But the most important part are the quick keys, which are not only quick but also allow you to manipulate lists and regions in a straightforward, unified way. And of course, the above 3 feature groups also integrate with quick keys.